requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.6.0
scikit-learn>=1.3.0
hdbscan>=0.8.33
pyarrow>=12.0.0
PyYAML>=6.0.0
google-generativeai>=0.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

pysbd>=0.3.4
